backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
psycopg2-binary>=2.9,<3.0
celery>=5.4,<6.0
redis>=5.0,<6.0
django-cors-headers>=4.4,<5.0
gunicorn>=22.0,<23.0
whitenoise>=6.7,<7.0
django-environ>=0.11,<1.0
